+++ lua/core/options.lua
@@ -61,4 +61,10 @@ M = {
    },
 }
 
+-- Improved Neovim's `ripgrep` defaults for `:grep`, if available
+if vim.fn.executable("rg") == 1 then
+   vim.opt.grepprg = "rg --vimgrep --smart-case --hidden"
+   vim.opt.grepformat = "%f:%l:%c:%m"
+end
+
 return M